Feat: added account helper
